Adds initial PanZoom and PanZoomDragUI classes
These are not yet complete. These are the current prototype states
from testing elsewhere. Still need to add the ability to call other
code when events change (specifically in the Drag class) so we can
trigger other UI items as needed.

These are functional, but we haven't started to consider developer
experience yet.